requirements.txt
bs4==0.0.2
certstream==1.12
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-mssql-backend==2.8.1
djangorestframework==3.14.0
gunicorn==21.2.0
ioc-fanger==4.2.1
ipwhois==1.2.0
Levenshtein==0.23.0
mssql-django==1.4
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.0
psycopg2-binary==2.9.9
pulsedive==0.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
tldextract==5.1.1
Unidecode==1.3.8
waitress==2.1.2
whitenoise==6.6.0